Move export logs mi to controls
